+++ KennelAPI/Controllers/UserController.cs
@@ -5,10 +5,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Common.Interfaces;
+using MongoPersistence.Entities;
+using AutoMapper;
+using Common.Entities;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace KennelAPI.Controllers
 {
-   /* [Route("api/users")]
+    [Route("api/users")]
     public class UserController : Controller
     {
         private readonly IUserRepository _userRepository;
@@ -19,9 +23,27 @@ namespace KennelAPI.Controllers
         }
 
         [HttpGet("{userId}")]
-        public IActionResult GetUser(int userId)
+        public async Task<IActionResult> GetUser(string userId)
         {
-            var user = _userRepository.GetUser(userId);
+            var bearerToken = Request.Headers["Authorization"];
+            var actualToken = bearerToken.FirstOrDefault();
+            var tokenSplit = actualToken?.Split(' '); //nullable operator
+
+            if (tokenSplit == null || tokenSplit.Length < 2)
+            {
+                //how to check for length < 2
+                return new UnauthorizedObjectResult(null);
+            }
+            var jwtToken = new JwtSecurityToken(actualToken.Split(' ')[1]);
+            var claims = jwtToken.Claims;
+            var userIdFromToken = claims.FirstOrDefault(c => c.Type == "userId");
+
+            if (userIdFromToken == null || userIdFromToken.Value != userId)
+            {
+                return new UnauthorizedObjectResult(null);
+            }
+
+            var user = await _userRepository.GetUser(userId);
 
             if (user == null)
             {
@@ -32,8 +54,9 @@ namespace KennelAPI.Controllers
         }
 
         [HttpPost()]
-        public IActionResult AddUser([FromBody] UserDtoCreation userDtoCreation)
+        public async Task<IActionResult> AddUser([FromBody] UserDtoCreation userDtoCreation)
         {
+            //how to setup token
             if (userDtoCreation == null)
             {
                 return NotFound();
@@ -44,34 +67,57 @@ namespace KennelAPI.Controllers
                 return BadRequest(ModelState);
             }
 
+            IUserEntity userEntity;
+
             try
             {
-                _userRepository.AddUser(userDtoCreation.Name, userDtoCreation.Phone, userDtoCreation.Email);
+                userEntity = Mapper.Map<UserEntity>(userDtoCreation);
+                userEntity.UserID = Guid.NewGuid().ToString();
+                await _userRepository.AddUser(userEntity);
             }
             catch (Exception)
             {
-                return StatusCode(500, "500 Bad Request");
+                return StatusCode(500, "500 Internal Server Error");
             }
 
             return Ok(userDtoCreation);
         }
 
         [HttpPut("{userId}")]
-        public IActionResult UpdateUsers(int userId, [FromBody] UserDtoUpdate userDtoUpdate)
+        public async Task<IActionResult> UpdateUsers(string userId, [FromBody] UserDtoUpdate userDtoUpdate)
         {
-            if (userId == null || userDtoUpdate == null)
+            var bearerToken = Request.Headers["Authorization"];
+            var actualToken = bearerToken.FirstOrDefault();
+            var tokenSplit = actualToken?.Split(' '); //nullable operator
+
+            if (tokenSplit == null || tokenSplit.Length < 2)
             {
-                return BadRequest();
+                //how to check for length < 2
+                return new UnauthorizedObjectResult(null);
+            }
+            var jwtToken = new JwtSecurityToken(actualToken.Split(' ')[1]);
+            var claims = jwtToken.Claims;
+            var userIdFromToken = claims.FirstOrDefault(c => c.Type == "userId");
+
+            if (userIdFromToken == null || userIdFromToken.Value != userId)
+            {
+                return new UnauthorizedObjectResult(null);
             }
 
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid || userDtoUpdate == null)
             {
-                return BadRequest();
+                return BadRequest(ModelState);
             }
 
-            var userToUpdate = _userRepository.GetUser(userId).Clone();
+            /*if(!ModelState.IsValid)
+            {
+                return BadRequest();
+            }*/
 
-            if(userToUpdate == null)
+            var userToUpdate = await _userRepository.GetUser(userId);
+            var userEntity = userToUpdate?.Clone();
+
+            if (userToUpdate == null)
             {
                 return NotFound();
             }
@@ -96,23 +142,42 @@ namespace KennelAPI.Controllers
         }
 
         [HttpDelete("{userId}")]
-        public IActionResult DeleteUser(int userId)
+        public async Task<IActionResult> DeleteUser(string userId)
         {
-            if(userId == null)
+            var bearerToken = Request.Headers["Authorization"];
+            var actualToken = bearerToken.FirstOrDefault();
+            var tokenSplit = actualToken?.Split(' '); //nullable operator
+
+            if (tokenSplit == null || tokenSplit.Length < 2)
             {
-                return BadRequest();
+                //how to check for length < 2
+                return new UnauthorizedObjectResult(null);
+            }
+            var jwtToken = new JwtSecurityToken(actualToken.Split(' ')[1]);
+            var claims = jwtToken.Claims;
+            var userIdFromToken = claims.FirstOrDefault(c => c.Type == "userId");
+
+            if (userIdFromToken == null || userIdFromToken.Value != userId)
+            {
+                return new UnauthorizedObjectResult(null);
             }
 
-            var userToDelete =_userRepository.GetUser(userId);
+            if (userId == null)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var userToDelete = await _userRepository.GetUser(userId);
 
             if(userToDelete == null)
             {
-                return NotFound();
+                //return NotFound();
+                return new UnauthorizedObjectResult(null);
             }
 
             _userRepository.DeleteUser(userToDelete);
 
             return NoContent();
         }
-    }*/
+    }
 }
